+++ src/ezdxf/entities/insert.py
@@ -36,6 +36,8 @@ if TYPE_CHECKING:
 
 __all__ = ['Insert']
 
+ABS_TOL = 1e-9
+
 # Multi-INSERT has subclass id AcDbMInsertBlock
 acdb_block_reference = DefSubclass('AcDbBlockReference', {
     'attribs_follow': DXFAttr(66, default=0, optional=True),
@@ -374,33 +376,28 @@ class Insert(LinkedEntities):
         """
 
         dxf = self.dxf
-        m1 = self.matrix44()
+        ocs = self.ocs()
 
-        # Transform scaled source axis into target coordinate system
-        ux, uy, uz = m.transform_directions((m1.ux, m1.uy, m1.uz))
+        # Transform source OCS axis into the target coordinate system:
+        ux, uy, uz = m.transform_directions((ocs.ux, ocs.uy, ocs.uz))
 
-        # Get new scaling factors, all are positive:
-        # z-axis is the real new z-axis, no reflection required
-        # x-axis is the real new x-axis, no reflection required
-        # y-axis - reflection is detected below
-        z_scale = uz.magnitude
-        x_scale = ux.magnitude
-        y_scale = uy.magnitude
+        # Calculate new axis scaling factors:
+        x_scale = ux.magnitude * dxf.xscale
+        y_scale = uy.magnitude * dxf.yscale
+        z_scale = uz.magnitude * dxf.zscale
 
-        # check for orthogonal x-, y- and z-axis
         ux = ux.normalize()
         uy = uy.normalize()
         uz = uz.normalize()
-        if not (math.isclose(ux.dot(uz), 0.0, abs_tol=1e-9) and
-                math.isclose(ux.dot(uy), 0.0, abs_tol=1e-9) and
-                math.isclose(uz.dot(uy), 0.0, abs_tol=1e-9)):
+        # check for orthogonal x-, y- and z-axis
+        if (abs(ux.dot(uz)) > ABS_TOL or abs(ux.dot(uy)) > ABS_TOL or
+                abs(uz.dot(uy)) > ABS_TOL):
             raise InsertTransformationError(NON_ORTHO_MSG)
 
-        # expected y-axis for an orthogonal right handed coordinate system
+        # expected y-axis for an orthogonal right handed coordinate system:
         expected_uy = uz.cross(ux)
-        if expected_uy.isclose(-uy, abs_tol=1e-9):
-            # transformed y-axis points into opposite direction of the expected
-            # y-axis:
+        if not expected_uy.isclose(uy, abs_tol=ABS_TOL):
+            # new y-axis points into opposite direction:
             y_scale = -y_scale
 
         ocs = OCSTransform.from_ocs(OCS(dxf.extrusion), OCS(uz), m)
@@ -443,10 +440,11 @@ class Insert(LinkedEntities):
         ux = Vec3(ocs.to_wcs(X_AXIS))
         uy = Vec3(ocs.to_wcs(Y_AXIS))
         m = Matrix44.ucs(ux=ux * sx, uy=uy * sy, uz=extrusion * sz)
+        # same as Matrix44.ucs(ux, uy, extrusion) * Matrix44.scale(sx, sy, sz)
 
         angle = math.radians(dxf.rotation)
-        if angle != 0.0:
-            m = Matrix44.chain(m, Matrix44.axis_rotate(extrusion, angle))
+        if angle:
+            m *= Matrix44.axis_rotate(extrusion, angle)
 
         insert = ocs.to_wcs(dxf.get('insert', Vec3()))